[Refactor Seeder] refactor seeder code block into seperate reusable function
