+++ src/database/seeders/seeder.service.ts
@@ -2,8 +2,8 @@ import { Injectable, Logger, OnApplicationBootstrap } from '@nestjs/common';
 import PokemonEntity from '../entities/pokemon.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import * as XLSX from 'xlsx';
 import { PokemonEntityKeys } from '../entities/types';
+import readXLSXFile from 'src/utils/read-xlsx-file';
 
 @Injectable()
 export class SeederService implements OnApplicationBootstrap {
@@ -14,30 +14,11 @@ export class SeederService implements OnApplicationBootstrap {
   ) {}
 
   async create() {
-    const pokemons = this.read_data('dist/database/seeders/PokemonGo.xlsx');
+    const pokemons = readXLSXFile<PokemonEntity>(
+      'dist/database/seeders/PokemonGo.xlsx',
+      () => new PokemonEntity(),
+    );
     return pokemons.map(async (p) => await p.save());
-  }
-
-  read_data(filePath: string): Array<PokemonEntity> {
-    const workbook = XLSX.readFile(filePath);
-    const sheetName = workbook.SheetNames[0];
-    const worksheet = workbook.Sheets[sheetName];
-    const rows: Array<Array<never>> = XLSX.utils.sheet_to_json(worksheet, {
-      header: 1,
-    });
-
-    const headers = rows.shift() as string[];
-
-    const data = rows.map((row) => {
-      const pokemonEntity = new PokemonEntity();
-      for (let index = 0; index < row.length; index++) {
-        const header = headers[index] as PokemonEntityKeys;
-        pokemonEntity[header] =
-          row[index] === '' ? (null as never) : row[index];
-      }
-      return pokemonEntity;
-    });
-    return data;
   }
 
   async seed() {
